services/ingestion/requirements.txt
kiteconnect==4.2.0
pydantic==2.5.0
pika==1.3.2
redis==5.0.1
structlog==24.1.0
python-dotenv==1.0.0
